chore : 절대경로 설정 수정 및 컴포넌트 페이지 하위로 변경
