+++ src/pages/extension/components/Extension1/styles.ts
