+++ src/pages/extension/components/Extension2/styles.ts
