+++ src/pages/extension/index.tsx
@@ -1,8 +1,8 @@
-import Extension1 from "@/components/Extension/Extension1";
-import Extension2 from "@/components/Extension/Extension2";
-import Extension3 from "@/components/Extension/Extension3";
-import Extension4 from "@/components/Extension/Extension4";
-import Extension5 from "@/components/Extension/Extension5";
+import Extension1 from "@/pages/extension/components/Extension1";
+import Extension2 from "@/pages/extension/components/Extension2";
+import Extension3 from "@/pages/extension/components/Extension3";
+import Extension4 from "@/pages/extension/components/Extension4";
+import Extension5 from "@/pages/extension/components/Extension5";
 import { ExtensionPageContainer } from "./styles";
 
 const ExtensionPage = () => {
