+++ src/pages/price/components/FAQ.tsx
