+++ src/pages/price/components/Footer.tsx
