+++ tsconfig.app.json
@@ -23,8 +23,7 @@
         /* 절대 경로 */
         "baseUrl": ".",
         "paths": {
-            "@/*": ["src/*"],
-            "@components/*": ["src/components/*"]
+            "@/*": ["src/*"]
         }
     },
     "include": ["src", "**/*.ts", "**/*.tsx"]
