+++ vite.config.ts
@@ -14,10 +14,7 @@ export default defineConfig({
         }),
     ],
     resolve: {
-        alias: [
-            { find: "@", replacement: "/src" },
-            { find: "@components", replacement: "/src/components" },
-        ],
+        alias: [{ find: "@", replacement: "/src" }],
     },
     envDir: "./pocket-prompt-frontend-envs/",
 });
